some progress but wrong rasterisation still

+++ src/Shaders/highThresh.frag
@@ -1,8 +1,8 @@
 #version 460
 precision highp float;
 
-layout (location = 0) out uint outMask;
+layout (location = 0) out float outMask;
 
 void main() {
-	outMask = 1;
+	outMask = 1.0;
 }
